Fixes to the working version of the Environment & Enable batch testing
I noticed that the x-axis of the graphs in Wandb was not consistent, that is fixed in this commit. Everything was caused by logging in several parts, each time wandb.log is called it assumes a new time step. Batch testing is enabled too considering some additional features in the environment and microgrid class.